refactor(transaction-pay-controller): split PolymarketStrategy into quotes + submit modules
Reorganises strategy/polymarket-bridge/ to match the layout of the other
strategies (across, relay) and removes the dead code from the prior
experimental phases:

- PolymarketBridgeStrategy class renamed to PolymarketStrategy, file
  renamed to match. Class shrinks from 945 to 43 lines and delegates to
  the two new modules.
- New polymarket-quotes.ts owns Relay quote fetch + TransactionPayQuote
  normalisation.
- New polymarket-submit.ts owns the deposit-wallet batched approve +
  unwrap to Relay's deposit address, Relay status polling, the USDC.e
  sweep (approve + wrap back to pUSD), the deposit-wallet batch transport
  with EIP-712 signing + relayer submission, and the 'wallet busy' retry.
- New polymarket-calldata.ts owns the ABI encoders for approve, unwrap,
  wrap, and the ERC-20 transfer-recipient extractor.
- Removed bridge-api.ts (no longer needed - Polymarket's own bridge API
  is no longer called).
- Removed withdraw.ts (its surface moved into polymarket-submit.ts as
  submitDepositWalletBatch with retry collapsed into the same function).
- Removed dead flags USE_RELAY_BRIDGE, USE_RELAY_DEPOSIT_ADDRESS,
  FORCE_SKIP_RELAY_POLL and the legacy execute branches they gated.
- PolymarketBridgeQuote slimmed to { relayQuote } - the wrapper exists
  only so the strategy can carry a typed quote through the controller.
