+++ packages/transaction-pay-controller/src/utils/strategy.ts
@@ -2,7 +2,7 @@ import { TransactionPayStrategy } from '../constants';
 import { AcrossStrategy } from '../strategy/across/AcrossStrategy';
 import { BridgeStrategy } from '../strategy/bridge/BridgeStrategy';
 import { FiatStrategy } from '../strategy/fiat/FiatStrategy';
-import { PolymarketBridgeStrategy } from '../strategy/polymarket-bridge/PolymarketBridgeStrategy';
+import { PolymarketStrategy } from '../strategy/polymarket-bridge/PolymarketStrategy';
 import { RelayStrategy } from '../strategy/relay/RelayStrategy';
 import { TestStrategy } from '../strategy/test/TestStrategy';
 import type {
@@ -39,7 +39,7 @@ export function getStrategyByName(
       return new FiatStrategy() as never;
 
     case TransactionPayStrategy.PolymarketBridge:
-      return new PolymarketBridgeStrategy() as never;
+      return new PolymarketStrategy() as never;
 
     case TransactionPayStrategy.Test:
       return new TestStrategy() as never;
